CCS-3942: combine assembly and module code
